Worker/requirements.txt
sentence-transformers
sqlalchemy
psycopg[binary]
pgvector
python-dotenv
